+++ src/main/java/fr/inria/corese/gui/core/service/ReasoningProfile.java
@@ -6,7 +6,7 @@ package fr.inria.corese.gui.core.service;
 public enum ReasoningProfile {
 	RDFS("RDFS entailments", "urn:corese:inference:rdfs"), OWL_RL("OWL RL entailments",
 			"urn:corese:inference:owlrl"), OWL_RL_LITE("OWL RL Lite entailments",
-					"urn:corese:inference:owlrl-lite"), OWL_RL_EXT("OWL RL Extended entailments",
+					"urn:corese:inference:owlrl-lite"), OWL_RL_EXT("OWL RL Ext entailments",
 							"urn:corese:inference:owlrl-ext");
 
 	private final String label;